requirements.txt
python-whois>=0.8.0
dnspython>=2.3.0
shodan>=1.29.0
requests>=2.28.0